refactor(InputView): 오프닝 멘트 출력  OutputView->InputView로 이동
