+++ src/main/java/vendingmachine/ui/OutputView.java
@@ -1,9 +1,5 @@
 package vendingmachine.ui;
 
-import static vendingmachine.ui.MessageUtils.INPUT_MACHINE_MONEY;
-import static vendingmachine.ui.MessageUtils.INPUT_MERCHANDISE_INFO;
-import static vendingmachine.ui.MessageUtils.INPUT_USERS_MONEY;
-import static vendingmachine.ui.MessageUtils.INPUT_USERS_STOCK_CHOICE;
 import static vendingmachine.ui.MessageUtils.MACHINE_COIN_INFO;
 import static vendingmachine.ui.MessageUtils.MONEY_REMAINING;
 
@@ -23,10 +19,6 @@ public class OutputView {
         System.out.println(ERROR + " " + errorMsg);
     }
 
-    public void printMachineMoneyInputOpening(){
-        System.out.println(INPUT_MACHINE_MONEY.msg());
-    }
-
     public void printMachineCoinInfo(RemainingCoins remainingCoins) {
         System.out.println();
         System.out.println(MACHINE_COIN_INFO.msg());
@@ -36,24 +28,9 @@ public class OutputView {
         }
     }
 
-    public void printMerchandiseInfoOpening(){
-        System.out.println();
-        System.out.println(INPUT_MERCHANDISE_INFO.msg());
-    }
-
-    public void printUserMoneyInputOpening() {
-        System.out.println();
-        System.out.println(INPUT_USERS_MONEY.msg());
-    }
-
     public void printLeftovers(int money) {
         System.out.println();
         System.out.println(MONEY_REMAINING.msg() + money + WON);
-    }
-
-
-    public void printUsersStockChoiceOpening() {
-        System.out.println(INPUT_USERS_STOCK_CHOICE.msg());
     }
 
     public void printRemainingCoins(EnumMap<Coin, Integer> change) {
